chore: eliminate redundant code
