+++ tests/test_session_concurrency.py
@@ -7,24 +7,10 @@ import threading
 import platform
 import subprocess
 from chdb import session
+from utils import is_musl_linux
 
 
 test_concurrent_dir = ".tmp_test_session_concurrency"
-
-
-def is_musl_linux():
-    if platform.system() != "Linux":
-        return False
-    try:
-        result = subprocess.run(['ldd', '--version'], capture_output=True, text=True)
-        print(f"stdout: {result.stdout.lower()}")
-        print(f"stderr: {result.stderr.lower()}")
-        # Check both stdout and stderr for musl
-        output_text = (result.stdout + result.stderr).lower()
-        return 'musl' in output_text
-    except Exception as e:
-        print(f"Exception in is_musl_linux: {e}")
-        return False
 
 
 class TestSessionConcurrency(unittest.TestCase):
